+++ inspector/src/components/actionTabs/tabs/propertyGrids/animations/editorControls.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Observable } from "babylonjs/Misc/observable";
 import { PropertyChangedEvent } from "../../../../../components/propertyChangedEvent";
 import { Animation } from "babylonjs/Animations/animation";
-import { IconButtonLineComponent } from "../../../lines/iconButtonLineComponent";
+import { IconButtonLineComponent } from "../../../../../sharedUiComponents/lines/iconButtonLineComponent";
 import { NumericInputComponent } from "../../../../../sharedUiComponents/lines/numericInputComponent";
 import { AddAnimation } from "./addAnimation";
 import { AnimationListTree, SelectedCoordinate } from "./animationListTree";
