+++ inspector/src/components/actionTabs/tabs/propertyGrids/commonPropertyGridComponent.tsx
@@ -7,7 +7,7 @@ import { LineContainerComponent } from "../../lineContainerComponent";
 import { LockObject } from "./lockObject";
 import { GlobalState } from "../../../globalState";
 import { TextLineComponent } from '../../lines/textLineComponent';
-import { IndentedTextLineComponent } from '../../lines/indentedTextLineComponent';
+import { IndentedTextLineComponent } from '../../../../sharedUiComponents/lines/indentedTextLineComponent';
 
 interface ICommonPropertyGridComponentProps {
     globalState: GlobalState;
